requirements.txt
aiohttp
aiosignal
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-timeout
attrs
backcall
beautifulsoup4
bleach
blinker
certifi
cffi
charset-normalizer
click
comm
dataclasses-json
debugpy
decorator
defusedxml
entrypoints
exceptiongroup
executing
faiss-cpu
fastapi
fastjsonschema
Flask
Flask-Cors
fqdn
frozenlist
greenlet
h11
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
isoduration
itsdangerous
jedi
Jinja2
jsonpointer
jsonschema
jupyter-events
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab-pygments
langchain
langchain_openai
langchainplus-sdk
MarkupSafe
marshmallow
marshmallow-enum
matplotlib-inline
mistune
multidict
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numexpr
numpy
openai
openapi-schema-pydantic
overrides
packaging
pandocfilters
parso
pexpect
pickle5
pickleshare
platformdirs
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
pydantic
Pygments
PyPDF2
pyrsistent
python-dateutil
python-dotenv
python-json-logger
python-multipart
PyYAML
pyzmq
regex
requests
rfc3339-validator
rfc3986-validator
Send2Trash
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
tenacity
terminado
tiktoken
tinycss2
tornado
tqdm
traitlets
typing
typing-inspect
typing_extensions
uri-template
urllib3
uvicorn
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
yarl
youtube-transcript-api
zipp